refactor(spec): extract normalize-scales and normalize-coord
Move position/color scale rewrites and coordinate-system canonicalization
out of normalize.ts so layer/aes orchestration stays separate. Split the
M2 normalize tests into their own file. Public normalize() path is unchanged.
