+++ .changeset/spec-normalize-scales-coord-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split normalize scale and coordinate canonicalization into dedicated modules behind the existing normalize() entry. Public normalize/normalizeChannel exports and behavior are unchanged.
